move test scripts
