+++ lib/SableServer.js
@@ -19,7 +19,7 @@ class SableServer {
 		this.stdout = stdout || process.stdout;
 		this.stderr = stderr || process.stderr;
 		server.on('request', (req, res) => this.onRequest(req, res));
-		for (const middleware of middlewares) {
+		for (const middleware of this.middlewares) {
 			if (middleware.onStart) {
 				middleware.onStart(this);
 			}
