+++ test/browser.js
@@ -7,9 +7,9 @@ const {promisify} = require('util');
 const {Builder, By} = require('selenium-webdriver');
 const {Local} = require('browserstack-local');
 const {startServer} = require('..');
-const {capabilities} = require('./capabilities.js');
-const env = require('./env.js');
-const {markResult} = require('./markResult.js');
+const {capabilities} = require('./lib/capabilities.js');
+const env = require('./lib/env.js');
+const {markResult} = require('./lib/markResult.js');
 const timeout = 60000;
 const mkdtemp = promisify(fs.mkdtemp);
 const packageJSON = require('../package.json');
