+++ test/lib/capabilities.js
