+++ test/lib/markResult.js
